requirements.txt
dj-database-url==0.4.1
Django==2.2.16
django-forms-bootstrap==3.1.0
django-netfields==1.2.2
django-pglocks==1.0.2
djangorestframework==3.9.1
djangorestframework-csv==1.4.1
gunicorn==19.5.0
ipaddress==1.0.22
netaddr==0.7.19
psycopg2==2.8.4
python-dateutil==2.4.2
pytz==2018.6
six==1.11.0
whitenoise==3.3.0